+++ src/main/kotlin/carracing/domain/Car.kt
@@ -1,5 +1,7 @@
 package carracing.domain
 
+import kotlin.random.Random
+
 data class Car(
     val name: String,
 ) {
@@ -7,10 +9,22 @@ data class Car(
         private set
 
     fun move() {
-        position++
+        if (isMove()) {
+            position++
+        }
     }
 
     fun copy() = Car(name).apply {
         this.position = this@Car.position
+    }
+
+    private fun isMove(): Boolean {
+        val num = Random.nextInt(RANDOM_RANGE)
+        return num >= MOVE_VALUE
+    }
+
+    companion object {
+        private const val RANDOM_RANGE = 10
+        private const val MOVE_VALUE = 4
     }
 }
