--- src/main/kotlin/carracing/domain/RacingGameRule.kt
@@ -1,7 +1,0 @@
-package carracing.domain
-
-import carracing.domain.Car
-
-interface RacingGameRule {
-    fun execute(cars: List<Car>)
-}
