+++ src/test/kotlin/carracing/CarRacingValidatorTest.kt
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test
 
 class CarRacingValidatorTest {
     @Test
-    fun `자동차 이름 5자 초과`() {
+    fun `자동차 이름 5자 초과시 예외 발생`() {
         val carName = "carNameOver"
         assertThrows(IllegalArgumentException::class.java) {
             CarRacingValidator().validateCarName(carName)
