+++ src/test/kotlin/carracing/RandomGameTest.kt
@@ -1,7 +1,6 @@
 package carracing
 
 import carracing.domain.Car
-import carracing.domain.RandomGame
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
@@ -11,7 +10,7 @@ class RandomGameTest {
     fun `랜덤 게임 테스트`() {
         val name = "car"
         val cars = listOf(Car(name))
-        RandomGame().execute(cars)
+        cars.map { it.move() }
         assertThat(cars.first().position).isIn(0, 1)
         assertThat(cars.first().name).isEqualTo(name)
     }
